requirements.txt
numpy
sounddevice
scipy
requests
websocket-client
pynput
pyperclip
pyautogui
pygetwindow
packaging
pillow
matrix-nio
PyQt6
python-telegram-bot
pydub
